+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Fixed
 - Fixed a bug where API endpoints were returning cached responses for Live Preview requests. ([#143](https://github.com/craftcms/element-api/issues/143))
+- Fixed a bug where endpoints whose route params didn’t align with the endpoint arguments would return a misleading 404 message. ([#137](https://github.com/craftcms/element-api/issues/137))
 
 ## 2.7.0 - 2021-05-26